[dbt] move get_asset_spec to translator
